refactor: moving the Resource Group ID parser/validator into their own package
