--- azurerm/internal/services/resource/group_validator_test.go
@@ -1,92 +1,0 @@
-package resource
-
-import "testing"
-
-func TestValidateResourceGroupID(t *testing.T) {
-	cases := []struct {
-		ID     string
-		Errors int
-	}{
-		{
-			ID:     "",
-			Errors: 1,
-		},
-		{
-			ID:     "nonsense",
-			Errors: 1,
-		},
-		{
-			ID:     "/slash",
-			Errors: 1,
-		},
-		{
-			ID:     "/path/to/nothing",
-			Errors: 1,
-		},
-		{
-			ID:     "/subscriptions",
-			Errors: 1,
-		},
-		{
-			ID:     "/providers",
-			Errors: 1,
-		},
-		{
-			ID:     "/subscriptions/not-a-guid",
-			Errors: 0,
-		},
-		{
-			ID:     "/providers/test",
-			Errors: 0,
-		},
-		{
-			ID:     "/subscriptions/00000000-0000-0000-0000-00000000000/",
-			Errors: 0,
-		},
-		{
-			ID:     "/providers/provider.name/",
-			Errors: 0,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.ID, func(t *testing.T) {
-			_, errors := ValidateResourceGroupID(tc.ID, "test")
-
-			if len(errors) < tc.Errors {
-				t.Fatalf("Expected ValidateResourceGroupID to have %d not %d errors for %q", tc.Errors, len(errors), tc.ID)
-			}
-		})
-	}
-}
-
-func TestValidateResourceGroupIDOrEmpty(t *testing.T) {
-	cases := []struct {
-		ID     string
-		Errors int
-	}{
-		{
-			ID:     "",
-			Errors: 0,
-		},
-		{
-			ID:     "nonsense",
-			Errors: 1,
-		},
-		// as this function just calls TestAzureResourceGroupId lets not be as comprehensive
-		{
-			ID:     "/providers/provider.name/",
-			Errors: 0,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.ID, func(t *testing.T) {
-			_, errors := ValidateResourceGroupIDOrEmpty(tc.ID, "test")
-
-			if len(errors) < tc.Errors {
-				t.Fatalf("Expected TestAzureResourceGroupIdOrEmpty to have %d not %d errors for %q", tc.Errors, len(errors), tc.ID)
-			}
-		})
-	}
-}
