+++ azurerm/internal/services/resource/validate/group_test.go
@@ -1,0 +1,61 @@
+package validate
+
+import "testing"
+
+func TestValidateResourceGroupID(t *testing.T) {
+	cases := []struct {
+		ID     string
+		Errors int
+	}{
+		{
+			ID:     "",
+			Errors: 1,
+		},
+		{
+			ID:     "nonsense",
+			Errors: 1,
+		},
+		{
+			ID:     "/slash",
+			Errors: 1,
+		},
+		{
+			ID:     "/path/to/nothing",
+			Errors: 1,
+		},
+		{
+			ID:     "/subscriptions",
+			Errors: 1,
+		},
+		{
+			ID:     "/providers",
+			Errors: 1,
+		},
+		{
+			ID:     "/subscriptions/not-a-guid",
+			Errors: 0,
+		},
+		{
+			ID:     "/providers/test",
+			Errors: 0,
+		},
+		{
+			ID:     "/subscriptions/00000000-0000-0000-0000-00000000000/",
+			Errors: 0,
+		},
+		{
+			ID:     "/providers/provider.name/",
+			Errors: 0,
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.ID, func(t *testing.T) {
+			_, errors := ResourceGroupID(tc.ID, "test")
+
+			if len(errors) < tc.Errors {
+				t.Fatalf("Expected ResourceGroupID to have %d not %d errors for %q", tc.Errors, len(errors), tc.ID)
+			}
+		})
+	}
+}
